JIRA: SWLIB-230 Expose triplanar functions
ImageTests was also not producing functions in the right way. It
was just casting to an ExprFunc instead of correctly creating one
that was static.

Fixing this led to the other problem in the previous commit
where name lookup of the global function table was being used
i.e. so resolveFunc() provided functions would not be LLVM evaluatable