JIRA: SEEXPR-147 Fix if then else behavior and variable upconverting
- Create ExprVarEnvBuilder. The idea is it holds a list of variable
contexts. As functions code gen they can swtich this out. This is
necessary because we need to be able to update the scope even if we are
deep in the function. I.e. a merge of an ifthenElse changes the
variables that are available for all subsequent
- Marshall the ExprVarEnvBuilder instead of ExprVarEnv around all the
prep() functions
- Hold a ExprVarEnvBuilder (which owns the ExprVarEnv's) in Expression.
That way we can reset it and we don't leak memory
- Update the prep to check types in the proper way. We track all the
implicitly created merge variables that will need to be code gen'd
_ Update LLVM code gen
- Disable local functions for now
- Use std::unique_ptr's to make memory management easier
- make BasicTests. IfThenElse which test the variable scope behavior
(execution and type checking)