Fixes SEEXPR-215: improve vector performance
These changes look pedantic, but they're actually avoiding some data movement by not requiring a copy construction to do arithmetic instructions on vector objects. It's subtle, but gives roughly 40% speedup in math-heavy use. This was verified by testing with a test program that simulates the heavy use of cJitter in the WIR2 princess scene, and also by doing code analysis of the generated assembly (we now generate 6 less mov instructions).