build: move code to src/SeExpr2 and src/SeExpr2/UI to simplify build
Remove the hacky build steps where we create a shadow include directory
structure within the temporary build directory.

Rename the source directories to match the final include directory names
so that no magic needs to be done to use the headers the same way that
external consumers would use them.