refactor to allow VarBlock to assign to const pointers directly
These changes change VarBlock to allow for assignment to some const pointer
directly. This makes the intention and safety clear, and allows calling
applications to assign some output buffer other than another VarBlock
variable. A nice side-effect of this is that we require less const_casts
in SeExpr2 integrations using VarBlocks. For backwards compatability, the old
API on evalMultiple() that accepts some VarBlock offset is reimplemented.