implements DeferredVars
DeferredVars allow callers to specify lambas (or other deferred computation storage) to VarBlocks to define variables.

Because of implementation details on how the Interpreter and LLVMEvaluator work, this requires two new class definition, DeferredRef and DeferredVarRef. This is not ideal because they really don't do anything special. The problem is in how the Interpreter and LLVMEvaluator use RTTI to select code generation patterns. Future work should remove our dependence on RTTI for code generation